+++ src/test/java/card/CardTest.java
@@ -1,10 +1,7 @@
-package domain.card;
+package card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import card.Card;
-import card.CardNumber;
-import card.CardType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
