+++ src/test/java/game/GameResultTest.java
@@ -1,15 +1,14 @@
-package domain;
+package game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import card.Card;
 import card.CardNumber;
 import card.CardType;
-import game.GameResult;
-import participant.Dealer;
-import participant.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import participant.Dealer;
+import participant.Player;
 
 class GameResultTest {
